Some fixes concerning floating point numbers
git-svn-id: https://alchemy-os.googlecode.com/svn/trunk@76 2c98f5b3-ddac-29b7-7766-ceeede3b9166

+++ alchemy-midlet/src/alchemy/evm/ELibBuilder.java
@@ -104,7 +104,7 @@ public class ELibBuilder implements LibBuilder {
 					cpool[cindex] = new Float(data.readFloat());
 					break;
 				case 'd': //double
-					cpool[cindex] = new Float(data.readDouble());
+					cpool[cindex] = new Double(data.readDouble());
 					break;
 				case 'S': //string
 					cpool[cindex] = data.readUTF();
